step-agents/account-provisioner/requirements.txt
pyyaml>=6.0
firebase-admin>=6.0.0
google-cloud-storage>=2.0.0
anthropic>=0.40.0
requests>=2.31.0
